pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"
skip_gitignore = true
float_to_top = true
default_section = "THIRDPARTY"
known_first_party = "xradarsat2"

[tool.setuptools_scm]
local_scheme = "no-local-version"
fallback_version = "9999"

[project]
name = "xradarsat2"
authors = [{ name = "Yann Reynaud", email = "Yann.Reynaud.2@ifremer.fr" }]
license = { text = "MIT" }
description = "xarray Level-1 SAR file reader for radarSat2"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  'xmltodict',
  'numpy',
  "xarray>=2024.10.0",
  'rasterio',
  'rioxarray>=0.18.1',
  'dask',
  'affine',
  'regex',
  'pyyaml',
  'fsspec',
  'aiohttp',

]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/umr-lops/xradarsat2"
"Bug Tracker" = "https://github.com/umr-lops/xradarsat2/issues"

[tool.coverage.report]
show_missing = true
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING"]

[tool.ruff.lint]
ignore = [
  "E402",  # module level import not at top of file
  "E501",  # line too long - let black worry about that
  "E731",  # do not assign a lambda expression, use a def
  "UP038", # type union instead of tuple for isinstance etc
]
select = [
  "F",   # Pyflakes
  "E",   # Pycodestyle
  "I",   # isort
  "UP",  # Pyupgrade
  "TID", # flake8-tidy-imports
  "W",
]
extend-safe-fixes = [
  "TID252", # absolute imports
  "UP031",  # percent string interpolation
]
fixable = ["I", "TID252", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["xradarsat2"]
known-third-party = ["xarray", "toolz", "construct"]

[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"
